Module Wall

module Color : sig ... end

Definition of colors, taken from Gg

module Transform : sig ... end

Definition of affine transformation matrices.

Wall drawing model

* * Drawing in wall is achieved by intersecting a simple, infinite image with a * shape. * * This image is described by a 'a Paint.t value and is simple by nature: * - a single color, * - a few different kinds of gradients, * - some user-defined pattern or textures, as determined by 'a. * * The 'a depends on the renderer and in practice it will be a * Wall_texture.t, an abstraction over OpenGL texture. * * The shapes are made from a Path.t that is filled or stroked. * The path is list of points that are connected by straight or curved (bezier) * lines. * When filled, the path is interpreted as the contour of a surface and the * resulting image is this surface. * When stroked, the path is interpreted as one or more lines: an Outline.t * that describe the style of line rendering (thickness, square or round ends, * etc) is used transform the abstract lines into a surface.

module Paint : sig ... end

module Outline : sig ... end

module Path : sig ... end

module Texture = Wall_texture

module Typesetter : sig ... end

module Image : sig ... end

module Performance_counter : sig ... end

module Renderer : sig ... end

type color = Color.t

type transform = Transform.t

type outline = Outline.t

type path = Path.t

type texture = Texture.t

type image = Image.t

type renderer = Renderer.t

type 'texture paint = 'texture Paint.t

type 'input typesetter = 'input Typesetter.t

val pi : float