Module Wall.Renderer

type t

A renderer allocates the OpenGL resources that are necessary to render contents.

val create : ?⁠antialias:bool -> ?⁠stencil_strokes:bool -> unit -> t

create ~antialias creates a new drawing context. antialias determines whether antialiasing is on or off, though it is strongly recommended to turn it on.

val delete : t -> unit

Calling delete t releases all the resources associated to the drawing context t. It is incorrect to use this context after the call.

A context can retain a lot of memory, so it is good practice to release it if you are no longer going to use it.

val render : t -> ?⁠performance_counter:Performance_counter.t -> width:float -> height:float -> Image.t -> unit